requirements.txt
google-generativeai
python-telegram-bot[webhooks]
Flask
python-dotenv
PyMuPDF
Pillow
